% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clckhs.R, R/dplyr.R
\docType{package}
\name{clckhs}
\alias{clckhs}
\alias{clickhouse}
\alias{clckhs-package}
\alias{copy_to.src_clickhouse}
\title{clckhs}
\usage{
\method{copy_to}{src_clickhouse}(dest, df, name = deparse(substitute(df)),
  overwrite = FALSE, ...)
}
\value{
a \code{tbl} object in the remote source
}
\description{
Provides Access to a clckhs Database
}
\seealso{
\code{\link[dplyr]{copy_to}}
}
