% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseConnection.R
\docType{class}
\name{ClickhouseConnection-class}
\alias{ClickhouseConnection-class}
\alias{show,ClickhouseConnection-method}
\alias{dbGetInfo,ClickhouseConnection-method}
\alias{dbIsValid,ClickhouseConnection-method}
\alias{dbDisconnect,ClickhouseConnection-method}
\alias{dbGetQuery,ClickhouseConnection,character-method}
\alias{dbSendQuery,ClickhouseConnection,character-method}
\alias{dbSendStatement,ClickhouseConnection,character-method}
\alias{dbExecute,ClickhouseConnection,character-method}
\alias{dbListTables,ClickhouseConnection-method}
\alias{dbExistsTable,ClickhouseConnection,character-method}
\alias{dbListFields,ClickhouseConnection,character-method}
\alias{dbReadTable,ClickhouseConnection,character-method}
\alias{dbRemoveTable,ClickhouseConnection,character-method}
\alias{dbCreateTable,ClickhouseConnection-method}
\alias{dbAppendTable,ClickhouseConnection-method}
\alias{dbWriteTable,ClickhouseConnection,character-method}
\alias{dbDataType,ClickhouseConnection-method}
\alias{dbQuoteIdentifier,ClickhouseConnection,character-method}
\alias{dbQuoteIdentifier,ClickhouseConnection,SQL-method}
\alias{dbQuoteString,ClickhouseConnection,character-method}
\alias{dbQuoteString,ClickhouseConnection,SQL-method}
\alias{dbBegin,ClickhouseConnection-method}
\alias{dbCommit,ClickhouseConnection-method}
\alias{dbRollback,ClickhouseConnection-method}
\title{Class ClickhouseConnection}
\usage{
\S4method{show}{ClickhouseConnection}(object)

\S4method{dbGetInfo}{ClickhouseConnection}(dbObj, ...)

\S4method{dbIsValid}{ClickhouseConnection}(dbObj, ...)

\S4method{dbDisconnect}{ClickhouseConnection}(conn, ...)

\S4method{dbGetQuery}{ClickhouseConnection,character}(conn, statement, ...)

\S4method{dbSendQuery}{ClickhouseConnection,character}(conn, statement, ...)

\S4method{dbSendStatement}{ClickhouseConnection,character}(conn, statement, ...)

\S4method{dbExecute}{ClickhouseConnection,character}(conn, statement, ...)

\S4method{dbListTables}{ClickhouseConnection}(conn, ...)

\S4method{dbExistsTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbListFields}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbReadTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbCreateTable}{ClickhouseConnection}(
  conn,
  name,
  fields,
  ...,
  engine = "MergeTree() ORDER BY tuple()",
  row.names = NULL,
  temporary = FALSE
)

\S4method{dbAppendTable}{ClickhouseConnection}(conn, name, value, ..., row.names = NULL)

\S4method{dbWriteTable}{ClickhouseConnection,character}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  engine = "MergeTree() ORDER BY tuple()",
  row.names = NA,
  field.types = NULL,
  ...
)

\S4method{dbDataType}{ClickhouseConnection}(dbObj, obj, ...)

\S4method{dbQuoteIdentifier}{ClickhouseConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{ClickhouseConnection,SQL}(conn, x, ...)

\S4method{dbQuoteString}{ClickhouseConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{ClickhouseConnection,SQL}(conn, x, ...)

\S4method{dbBegin}{ClickhouseConnection}(conn, ...)

\S4method{dbCommit}{ClickhouseConnection}(conn, ...)

\S4method{dbRollback}{ClickhouseConnection}(conn, ...)
}
\arguments{
\item{engine}{the ClickHouse table engine to use when creating a table.}
}
\description{
\code{ClickhouseConnection} objects are created by \code{\link[DBI:dbConnect]{DBI::dbConnect()}} and wrap a
transport (Arrow Flight SQL or HTTP). All DBI methods delegate to the
transport seam (see \code{transport.R}); the class exists so that dbplyr can
dispatch ClickHouse-specific SQL on it.
}
\keyword{internal}
