% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseDriver.R
\docType{class}
\name{ClickhouseDriver-class}
\alias{ClickhouseDriver-class}
\alias{clickhouse}
\alias{show,ClickhouseDriver-method}
\alias{dbGetInfo,ClickhouseDriver-method}
\alias{dbIsValid,ClickhouseDriver-method}
\alias{dbUnloadDriver,ClickhouseDriver-method}
\alias{dbConnect,ClickhouseDriver-method}
\alias{dbDataType,ClickhouseDriver-method}
\title{Class ClickhouseDriver}
\usage{
clickhouse()

\S4method{show}{ClickhouseDriver}(object)

\S4method{dbGetInfo}{ClickhouseDriver}(dbObj, ...)

\S4method{dbIsValid}{ClickhouseDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{ClickhouseDriver}(drv, ...)

\S4method{dbConnect}{ClickhouseDriver}(
  drv,
  host = "localhost",
  port = NULL,
  dbname = "default",
  user = "default",
  password = "",
  transport = c("flightsql", "http"),
  tls = FALSE,
  config_paths = default_config_paths(),
  ...
)

\S4method{dbDataType}{ClickhouseDriver}(dbObj, obj, ...)
}
\arguments{
\item{...}{passed to the transport. Also accepts deprecated arguments
(\code{compression}, \code{Int64}, \code{toUTF8}, \code{db}) for backward compatibility.}

\item{drv}{a \linkS4class{ClickhouseDriver} object created by \code{\link[=clickhouse]{clickhouse()}}.}

\item{host}{host on which the database is running.}

\item{port}{port to connect to. Defaults to the transport's standard port
(9090 for Flight SQL, 8123/8443 for HTTP).}

\item{dbname}{name of the default database.}

\item{user}{name of the user to connect as.}

\item{password}{the user's password.}

\item{transport}{communication backend: \code{"flightsql"} (Arrow Flight SQL via
ADBC, the default) or \code{"http"} (HTTP interface with the ArrowStream
format). See \link{RClickhouse-package}.}

\item{tls}{use a TLS-encrypted connection.}

\item{config_paths}{paths searched for YAML configuration files; earlier
paths take precedence.}
}
\value{
A \linkS4class{ClickhouseConnection} object.
}
\description{
This driver never needs to be unloaded and hence \code{dbUnloadDriver()} is a
no-op.
}
\examples{
library(DBI)
RClickhouse::clickhouse()
\dontrun{
conn <- dbConnect(RClickhouse::clickhouse(), host = "localhost")
conn <- dbConnect(RClickhouse::clickhouse(), host = "localhost",
                  port = 8123, transport = "http")
}
}
\keyword{internal}
