% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseResult.R
\docType{class}
\name{ClickhouseResult-class}
\alias{ClickhouseResult-class}
\alias{dbFetch,ClickhouseResult-method}
\alias{dbClearResult,ClickhouseResult-method}
\alias{dbHasCompleted,ClickhouseResult-method}
\alias{dbGetStatement,ClickhouseResult-method}
\alias{dbIsValid,ClickhouseResult-method}
\alias{dbGetRowCount,ClickhouseResult-method}
\alias{dbGetRowsAffected,ClickhouseResult-method}
\alias{dbColumnInfo,ClickhouseResult-method}
\title{Class ClickhouseResult}
\usage{
\S4method{dbFetch}{ClickhouseResult}(res, n = -1, ...)

\S4method{dbClearResult}{ClickhouseResult}(res, ...)

\S4method{dbHasCompleted}{ClickhouseResult}(res, ...)

\S4method{dbGetStatement}{ClickhouseResult}(res, ...)

\S4method{dbIsValid}{ClickhouseResult}(dbObj, ...)

\S4method{dbGetRowCount}{ClickhouseResult}(res, ...)

\S4method{dbGetRowsAffected}{ClickhouseResult}(res, ...)

\S4method{dbColumnInfo}{ClickhouseResult}(res, ...)
}
\description{
Encapsulates the result of a statement. Because both transports return the
full Arrow result eagerly, a \code{ClickhouseResult} holds the materialised
data.frame and a cursor; \code{\link[=dbFetch]{dbFetch()}} slices from it.
}
\keyword{internal}
