% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-helpers.R
\name{fix_dbplyr}
\alias{fix_dbplyr}
\title{Restore dbplyr's default function-name handling.}
\usage{
fix_dbplyr()
}
\value{
No return value, called for side effects.
}
\description{
Undoes \code{\link[=dbplyr_case_sensitive]{dbplyr_case_sensitive()}}.
}
