% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RClickhouse-package.R
\docType{package}
\name{RClickhouse-package}
\alias{RClickhouse}
\alias{RClickhouse-package}
\title{RClickhouse: a DBI interface to ClickHouse over Apache Arrow}
\description{
RClickhouse provides a \link[DBI:DBI-package]{DBI}-compliant interface to the
\href{https://clickhouse.com/}{ClickHouse} database. All communication uses
Apache Arrow as the wire format, so type conversion is handled natively on
both ends and the package contains no compiled code.
}
\details{
Two transports are available, selected with the \code{transport} argument of
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}:
\itemize{
\item \code{"flightsql"} (default): the native
\href{https://clickhouse.com/docs/interfaces/arrowflight}{Arrow Flight SQL}
protocol via ADBC. Requires the \code{adbcflightsql} and \code{adbcdrivermanager}
packages and a ClickHouse server with \code{arrowflight_port} enabled.
\item \code{"http"}: the ClickHouse HTTP interface using the \code{ArrowStream} format.
Pure R (\code{httr2} + \code{nanoarrow}), works against any ClickHouse server,
including ClickHouse Cloud.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IMSMWU/RClickhouse}
  \item Report bugs at \url{https://github.com/IMSMWU/RClickhouse/issues}
}

}
\author{
\strong{Maintainer}: Christian Hotz-Behofsits \email{christian.hotz-behofsits@wu.ac.at}

Authors:
\itemize{
  \item Daniel Winkler \email{daniel.winkler@wu.ac.at}
  \item Luca Rauchenberger \email{luca.rauchenberger@wu.ac.at}
  \item Peter Knaus \email{peter.knaus@wu.ac.at}
  \item Clemens Danninger \email{cran@b8000h.org}
  \item Daria Yudaeva \email{daria.yudaeva@wu.ac.at}
  \item Simon Stiebellehner \email{simon.stiebellehner@gmail.com}
}

}
\keyword{internal}
