% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-helpers.R
\name{sql_aggregate}
\alias{sql_aggregate}
\alias{sql_aggregate_2}
\title{adapted from dbplyr}
\usage{
sql_aggregate(f)

sql_aggregate_2(f)
}
\arguments{
\item{f}{SQL function name}
}
\description{
adapted from dbplyr

adapted from dbplyr
}
