% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseConnection.R
\docType{class}
\name{ClickhouseConnection-class}
\alias{ClickhouseConnection-class}
\alias{dbGetInfo,ClickhouseConnection-method}
\alias{dbDisconnect,ClickhouseConnection-method}
\title{Class ClickhouseConnection}
\usage{
\S4method{dbGetInfo}{ClickhouseConnection}(dbObj, ...)

\S4method{dbDisconnect}{ClickhouseConnection}(conn, ...)
}
\description{
\code{ClickhouseConnection.} objects are usually created by
\code{\link[DBI]{dbConnect}}

Close the database connection
}
\keyword{internal}
