% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseConnection.R
\docType{class}
\name{ClickhouseConnection-class}
\alias{ClickhouseConnection-class}
\alias{dbGetInfo,ClickhouseConnection-method}
\alias{dbIsValid,ClickhouseConnection-method}
\alias{dbListTables,ClickhouseConnection-method}
\alias{dbExistsTable,ClickhouseConnection,character-method}
\alias{dbReadTable,ClickhouseConnection,character-method}
\alias{dbRemoveTable,ClickhouseConnection,character-method}
\alias{dbListFields,ClickhouseConnection,character-method}
\alias{dbSendQuery,ClickhouseConnection,character-method}
\alias{dbDataType,ClickhouseConnection-method}
\alias{dbQuoteIdentifier,ClickhouseConnection,character-method}
\alias{dbBegin,ClickhouseConnection-method}
\alias{dbCommit,ClickhouseConnection-method}
\alias{dbRollback,ClickhouseConnection-method}
\alias{dbDisconnect,ClickhouseConnection-method}
\title{Class ClickhouseConnection}
\usage{
\S4method{dbGetInfo}{ClickhouseConnection}(dbObj, ...)

\S4method{dbIsValid}{ClickhouseConnection}(dbObj, ...)

\S4method{dbListTables}{ClickhouseConnection}(conn, ...)

\S4method{dbExistsTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbReadTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbListFields}{ClickhouseConnection,character}(conn, name, ...)

\S4method{dbSendQuery}{ClickhouseConnection,character}(conn, statement, ...)

\S4method{dbDataType}{ClickhouseConnection}(dbObj, obj, ...)

\S4method{dbQuoteIdentifier}{ClickhouseConnection,character}(conn, x, ...)

\S4method{dbBegin}{ClickhouseConnection}(conn, ...)

\S4method{dbCommit}{ClickhouseConnection}(conn, ...)

\S4method{dbRollback}{ClickhouseConnection}(conn, ...)

\S4method{dbDisconnect}{ClickhouseConnection}(conn, ...)
}
\description{
\code{ClickhouseConnection.} objects are usually created by
\code{\link[DBI]{dbConnect}}

Close the database connection
}
\keyword{internal}
