% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseDriver.R
\docType{class}
\name{ClickhouseDriver-class}
\alias{ClickhouseDriver-class}
\alias{dbIsValid,ClickhouseDriver-method}
\alias{dbUnloadDriver,ClickhouseDriver-method}
\alias{dbConnect,ClickhouseDriver-method}
\title{Class ClickhouseDriver}
\usage{
\S4method{dbIsValid}{ClickhouseDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{ClickhouseDriver}(drv, ...)

\S4method{dbConnect}{ClickhouseDriver}(drv, host = "localhost",
  port = 8123L, user = "default", password = "", ...)
}
\description{
This driver never needs to be unloaded and hence \code{dbUnload()} is a
null-op.
}
\keyword{internal}
