% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseResult.R
\docType{class}
\name{ClickhouseResult-class}
\alias{ClickhouseResult-class}
\alias{fetch,ClickhouseResult,numeric-method}
\alias{dbGetRowsAffected,ClickhouseResult-method}
\alias{dbClearResult,ClickhouseResult-method}
\alias{dbHasCompleted,ClickhouseResult-method}
\alias{dbIsValid,ClickhouseResult-method}
\title{Class ClickhouseResult}
\usage{
\S4method{fetch}{ClickhouseResult,numeric}(res, n = -1, ...)

\S4method{dbGetRowsAffected}{ClickhouseResult}(res, ...)

\S4method{dbClearResult}{ClickhouseResult}(res, ...)

\S4method{dbHasCompleted}{ClickhouseResult}(res, ...)

\S4method{dbIsValid}{ClickhouseResult}(dbObj, ...)
}
\description{
Clickhouse's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}
