% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickhouseResult.R
\docType{class}
\name{ClickhouseResult-class}
\alias{ClickhouseResult-class}
\alias{dbFetch,ClickhouseResult,numeric-method}
\alias{dbGetRowsAffected,ClickhouseResult-method}
\alias{dbClearResult,ClickhouseResult-method}
\alias{dbHasCompleted,ClickhouseResult-method}
\alias{dbIsValid,ClickhouseResult-method}
\alias{dbGetStatement,ClickhouseResult-method}
\alias{dbGetRowCount,ClickhouseResult-method}
\title{Class ClickhouseResult}
\usage{
\S4method{dbFetch}{ClickhouseResult,numeric}(res, n = -1, ...)

\S4method{dbGetRowsAffected}{ClickhouseResult}(res, ...)

\S4method{dbClearResult}{ClickhouseResult}(res, ...)

\S4method{dbHasCompleted}{ClickhouseResult}(res, ...)

\S4method{dbIsValid}{ClickhouseResult}(dbObj, ...)

\S4method{dbGetStatement}{ClickhouseResult}(res, ...)

\S4method{dbGetRowCount}{ClickhouseResult}(res, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Clickhouse's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}
