% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbplyr-helpers.R
\name{dbplyr_case_sensitive}
\alias{dbplyr_case_sensitive}
\title{Get dbplyr to work with Clickhouse}
\usage{
dbplyr_case_sensitive()
}
\description{
Functions passed to Clickhouse are case sensitive. By default dbplyr
converts functions that are not predefined to upper case. This function
changes that behavior to leave passed functions as they are.
}
